+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Misc:
 - **PHPMD** Handle syntax error as issue [#2201](https://github.com/sider/runners/pull/2201)
 - Add `linter.<id>.dependencies` option for npm tools [#2202](https://github.com/sider/runners/pull/2202)
 - **ktlint** Add issue ID for syntax error [#2206](https://github.com/sider/runners/pull/2206)
+- **Clang-Tidy** Better `apt-get install` failure [#2210](https://github.com/sider/runners/pull/2210)
 
 ## 0.45.0